Fix error messages that occur after degraded and progressing resources
Add "Requeue: true" to the reconcile return values to force the
reconciler to rerun after a resource is degraded or progressing.
This is preferred over throwing an error message in the return
value of the Reconcile function.

Additionally, fix the status message for the master daemonset, as
it was referring to the worker daemonset and not the master
daemonset.